feat: prompt for login any time a user omits API key
