+++ src/cmds/categories/create.ts
@@ -54,7 +54,7 @@ export default class CategoriesCreateCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { categoryType, title, key, version, preventDuplicates } = opts;
