+++ src/cmds/categories/index.ts
@@ -18,7 +18,7 @@ export default class CategoriesCommand extends Command {
   }
 
   async run(opts: CommandOptions<{}>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { key, version } = opts;
