+++ src/cmds/changelogs/index.ts
@@ -42,7 +42,7 @@ export default class ChangelogsCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { dryRun, folder, key } = opts;
