+++ src/cmds/changelogs/single.ts
@@ -41,7 +41,7 @@ export default class SingleChangelogCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { dryRun, filePath, key } = opts;
