+++ src/cmds/custompages/index.ts
@@ -42,7 +42,7 @@ export default class CustomPagesCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { dryRun, folder, key } = opts;
