+++ src/cmds/docs/edit.ts
@@ -45,7 +45,7 @@ export default class EditDocsCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>): Promise<undefined> {
-    super.run(opts);
+    await super.run(opts);
 
     const { slug, key, version } = opts;
