+++ src/cmds/docs/single.ts
@@ -42,7 +42,7 @@ export default class SingleDocCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { dryRun, filePath, key, version } = opts;
