+++ src/cmds/openapi/reduce.ts
@@ -45,7 +45,7 @@ export default class OpenAPIReduceCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { spec, workingDirectory } = opts;
