+++ src/cmds/openapi/validate.ts
@@ -38,7 +38,7 @@ export default class OpenAPIValidateCommand extends Command {
   }
 
   async run(opts: CommandOptions<Options>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { spec, workingDirectory } = opts;
