+++ src/cmds/versions/create.ts
@@ -47,7 +47,7 @@ export default class CreateVersionCommand extends Command {
   }
 
   async run(opts: CommandOptions<VersionCreateOptions>) {
-    super.run(opts);
+    await super.run(opts);
 
     let versionList;
     const { key, version, fork, codename, main, beta, isPublic } = opts;
