+++ src/cmds/versions/delete.ts
@@ -28,7 +28,7 @@ export default class DeleteVersionCommand extends Command {
   }
 
   async run(opts: CommandOptions<{}>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { key, version } = opts;
