+++ src/cmds/versions/index.ts
@@ -37,7 +37,7 @@ export default class VersionsCommand extends Command {
   }
 
   async run(opts: CommandOptions<{}>) {
-    super.run(opts);
+    await super.run(opts);
 
     const { key, version } = opts;
