+++ src/cmds/whoami.ts
@@ -20,7 +20,7 @@ export default class WhoAmICommand extends Command {
   }
 
   async run(opts: CommandOptions<{}>) {
-    super.run(opts);
+    await super.run(opts);
 
     if (!configStore.has('email') || !configStore.has('project')) {
       return Promise.reject(new Error(`Please login using \`${config.get('cli')} login\`.`));
